pyproject.toml
[tool.poetry]
name = "mail_ru_test_task"
version = "0.1.0"
description = "Test task for Mail.Ru"
authors = ["Roxe322 <kolomatskiy@bk.ru>"]

[tool.poetry.dependencies]
python = "^3.8"
pyyaml = "^5.3"
aioredis = "^1.3.1"
aiohttp = "^3.6.2"
pydantic = "^1.4"
marshmallow = "^3.4.0"
webargs = "^5.5.3"

[tool.poetry.dev-dependencies]
flake8 = "^3.7.9"
mypy = "^0.761"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
